move database to dataclass namespace
