+++ python/lib/database_lib/candidate_db.py
@@ -11,7 +11,7 @@ class CandidateDB:
     :Example:
 
         from lib.database_lib.candidate_db import CandidateDB
-        from lib.database import Database
+        from lib.dataclass.database import Database
 
         # database connection
         db = Database(config.mysql, verbose)
