+++ python/lib/database_lib/config.py
@@ -11,7 +11,7 @@ class Config:
     :Example:
 
         from lib.database_lib.config import Config
-        from lib.database import Database
+        from lib.dataclass.database import Database
 
         # database connection
         db = Database(config.mysql, verbose)
