+++ python/lib/database_lib/files.py
@@ -12,7 +12,7 @@ class Files:
     :Example:
 
         from lib.files import Files
-        from lib.database import Database
+        from lib.dataclass.database import Database
 
         # database connection
         db = Database(config.mysql, verbose)
