+++ tools/correct_lists_incorrectly_saved_in_parameter_file.py
@@ -2,7 +2,7 @@
 
 import os
 
-from lib.database import Database
+from lib.dataclass.database import Database
 from lib.lorisgetopt import LorisGetOpt
